完善上链代码

+++ ShareDeliveryInfo/ShareDeliveryInfo/mycontract/ShareDeliveryInfo.sol
@@ -3,18 +3,18 @@ pragma solidity >= 0.4.25;
 contract ShareDeliveryInfo {
     uint package_uid; //包裹的UID
 
-    mapping(uint => string) public delivery_path;//记录快递到达站点路径
+    mapping(uint64 => string) public delivery_path;//记录快递到达站点路径
 
     constructor() public {
     }
 
     //获取快递传送路径信息
-    function get(uint _pkg_uid) public view returns(string) {
+    function get(uint64 _pkg_uid) public view returns(string) {
         return delivery_path[_pkg_uid];
     }
 
     //快递到达新站点，更新信息
-    function set(uint _pkg_uid, string station) public {
+    function set(uint64 _pkg_uid, string station) public {
         delivery_path[_pkg_uid] = strConcat(delivery_path[_pkg_uid], "； 到达站点：");
         delivery_path[_pkg_uid] = strConcat(delivery_path[_pkg_uid], station);
     }
